+++ plib/controllers/SettingsController.php
@@ -20,6 +20,7 @@ class SettingsController extends pm_Controller_Action
 
         $this->view->tabs = [
             ['title' => $this->lmsg('tabs.dashboard'), 'action' => 'index', 'controller' => 'index'],
+            ['title' => $this->lmsg('tabs.filter'), 'action' => 'index', 'controller' => 'filter'],
             ['title' => $this->lmsg('tabs.settings'), 'action' => 'index', 'controller' => 'settings'],
         ];
     }